refactor: final polishing and documentation
