+++ packages/rs-json-schema-compatibility/src/error.rs
@@ -1,22 +1,31 @@
+use crate::CompatibilityRules;
 use json_patch::PatchOperation;
 use serde_json::Value;
 
+/// Compatibility validation errors.
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
+    /// An unexpected patch operation was encountered.
     #[error(transparent)]
     UnexpectedJsonPatchOperation(UnexpectedPatchOperationError),
+    /// The JSON Pointer path doesn't exist in the JSON value.
     #[error(transparent)]
     JsonPointerPathNotFound(JsonPointerPathNotFoundError),
+    /// There is no compatibility rules are defined for the encountered schema keyword
     #[error(transparent)]
     UnsupportedSchemaKeyword(UnsupportedSchemaKeywordError),
+    /// The JSON patch operation path is too small and doesn't contain keywords.
     #[error(transparent)]
     InvalidJsonPatchOperationPath(InvalidJsonPatchOperationPathError),
+    /// The [IsReplacementAllowedCallback] is not defined for the encountered keyword and [ReplaceOperation].
     #[error(transparent)]
-    UndefinedReplaceCallback(UndefinedReplaceCallbackError),
+    UndefinedReplaceCallback(UndefinedReplacementAllowedCallbackError),
+    /// The JSON value type is not as expected.
     #[error(transparent)]
     UnexpectedJsonValueType(UnexpectedJsonValueTypeError),
 }
 
+/// An unexpected patch operation was encountered.
 #[derive(thiserror::Error, Debug)]
 #[error("unexpected patch operation: {0}")]
 pub struct UnexpectedPatchOperationError(pub PatchOperation);
@@ -27,6 +36,7 @@ impl From<UnexpectedPatchOperationError> for Error {
     }
 }
 
+/// The JSON Pointer path doesn't exist in the JSON value.
 #[derive(thiserror::Error, Debug)]
 #[error("JSON Pointer path '{path}' doesn't exist in JSON value")]
 pub struct JsonPointerPathNotFoundError {
@@ -40,6 +50,7 @@ impl From<JsonPointerPathNotFoundError> for Error {
     }
 }
 
+/// There is no compatibility rules are defined for the encountered schema keyword
 #[derive(thiserror::Error, Debug)]
 #[error("schema keyword '{keyword}' at path '{path}' is not supported")]
 pub struct UnsupportedSchemaKeywordError {
@@ -53,6 +64,7 @@ impl From<UnsupportedSchemaKeywordError> for Error {
     }
 }
 
+/// The JSON patch operation path is too small and doesn't contain keywords.
 #[derive(thiserror::Error, Debug)]
 #[error("json patch operation path '{path}' is too small and doesn't contain keywords")]
 pub struct InvalidJsonPatchOperationPathError {
@@ -65,18 +77,21 @@ impl From<InvalidJsonPatchOperationPathError> for Error {
     }
 }
 
+/// The [IsReplacementAllowedCallback] is not defined for the encountered keyword and [ReplaceOperation].
 #[derive(thiserror::Error, Debug)]
-#[error("undefined allow replace callback for keyword '{keyword}'")]
-pub struct UndefinedReplaceCallbackError {
-    pub keyword: String,
+#[error("undefined allow replacement callback for path '{path}'")]
+pub struct UndefinedReplacementAllowedCallbackError {
+    pub path: String,
+    pub rules: CompatibilityRules,
 }
 
-impl From<UndefinedReplaceCallbackError> for Error {
-    fn from(e: UndefinedReplaceCallbackError) -> Self {
+impl From<UndefinedReplacementAllowedCallbackError> for Error {
+    fn from(e: UndefinedReplacementAllowedCallbackError) -> Self {
         Error::UndefinedReplaceCallback(e)
     }
 }
 
+/// The JSON value type is not as expected.
 #[derive(thiserror::Error, Debug)]
 #[error("expected JSON value type '{expected_type}', but got '{value}'")]
 pub struct UnexpectedJsonValueTypeError {
