--- packages/rs-json-schema-compatibility/src/keyword/mod.rs
@@ -1,6 +1,0 @@
-mod keyword_rule;
-mod rules;
-mod value;
-
-pub use keyword_rule::{KeywordRule, KeywordRuleExample, ReplaceCallback};
-pub use rules::KEYWORD_RULES;
