+++ packages/rs-json-schema-compatibility/src/lib.rs
@@ -1,9 +1,12 @@
 mod change;
 mod error;
-mod keyword;
+mod rules;
 mod validate;
 
 pub use crate::change::JsonSchemaChange;
-pub use crate::keyword::{KeywordRule, KeywordRuleExample, ReplaceCallback, KEYWORD_RULES};
+pub use crate::rules::{
+    CompatibilityRuleExample, CompatibilityRules, IsReplacementAllowedCallback,
+    KEYWORD_COMPATIBILITY_RULES,
+};
 pub use json_patch::{AddOperation, RemoveOperation, ReplaceOperation};
 pub use validate::{validate_schemas_compatibility, CompatibilityValidationResult};
