+++ packages/rs-json-schema-compatibility/src/rules/mod.rs
@@ -1,0 +1,8 @@
+mod compatibility_rules;
+mod rule_set;
+mod value;
+
+pub use compatibility_rules::{
+    CompatibilityRuleExample, CompatibilityRules, IsReplacementAllowedCallback,
+};
+pub use rule_set::KEYWORD_COMPATIBILITY_RULES;
